requirements.txt
langchain_google_genai
langchain
langgraph 
langchain-community 
langchain-openai 
duckduckgo-search
pyttsx3
langchain_google_genai
langchain
langgraph
langchain-community
langchain-openai
duckduckgo-search
pyttsx3
gtts
langchain_tavily
streamlit
langchain-groq
python-dotenv
langchain-xai
elevenlabs
uvicorn
pydantic-settings
fastapi
google-genai
pyttsx3
torch 
soundfile
omegaconf